fix: line chart user story
