+++ src/stories/containers/Finances/components/SectionPages/MakerDAOExpenseMetrics/useMakerDAOExpenseMetrics.tsx
@@ -1,4 +1,5 @@
 import { fetchAnalytics } from '@ses/containers/Finances/api/queries';
+import { buildExpenseMetricsLineChartSeries } from '@ses/containers/Finances/utils/utils';
 import { useThemeContext } from '@ses/core/context/ThemeContext';
 import { useMemo, useState } from 'react';
 import useSWRImmutable from 'swr/immutable';
@@ -56,63 +57,10 @@ export const useMakerDAOExpenseMetrics = (year: string) => {
     );
   };
 
-  const series: LineChartSeriesData[] = useMemo(() => {
-    const disabled = {
-      Budget: inactiveSeries.includes('Budget'),
-      Forecast: inactiveSeries.includes('Forecast'),
-      Actuals: inactiveSeries.includes('Actuals'),
-      'Net Expenses On-chain': inactiveSeries.includes('Net Expenses On-chain'),
-      'Net Expenses Off-chain': inactiveSeries.includes('Net Expenses Off-chain'),
-    };
-
-    return [
-      {
-        name: 'Budget',
-        data: disabled.Budget ? [] : data?.budget,
-        type: 'line',
-        itemStyle: {
-          color: disabled.Budget ? '#ccc' : isLight ? '#F99374' : '#F77249',
-        },
-        isVisible: !disabled.Budget,
-      },
-      {
-        name: 'Forecast',
-        data: disabled.Forecast ? [] : data?.forecast,
-        type: 'line',
-        itemStyle: {
-          color: disabled.Forecast ? '#ccc' : isLight ? '#447AFB' : '#447AFB',
-        },
-        isVisible: !disabled.Forecast,
-      },
-      {
-        name: 'Actuals',
-        data: disabled.Actuals ? [] : data?.actuals,
-        type: 'line',
-        itemStyle: {
-          color: disabled.Actuals ? '#ccc' : isLight ? '#2DC1B1' : '#1AAB9B',
-        },
-        isVisible: !disabled.Actuals,
-      },
-      {
-        name: 'Net Expenses On-chain',
-        data: disabled['Net Expenses On-chain'] ? [] : data?.onChain,
-        type: 'line',
-        itemStyle: {
-          color: disabled['Net Expenses On-chain'] ? '#ccc' : isLight ? '#FBCC5F' : '#FDC134',
-        },
-        isVisible: !disabled['Net Expenses On-chain'],
-      },
-      {
-        name: 'Net Expenses Off-chain',
-        data: disabled['Net Expenses Off-chain'] ? [] : data?.offChain,
-        type: 'line',
-        itemStyle: {
-          color: disabled['Net Expenses Off-chain'] ? '#ccc' : isLight ? '#7C6B95' : '#6C40AA',
-        },
-        isVisible: !disabled['Net Expenses Off-chain'],
-      },
-    ];
-  }, [data?.actuals, data?.budget, data?.forecast, data?.offChain, data?.onChain, inactiveSeries, isLight]);
+  const series: LineChartSeriesData[] = useMemo(
+    () => buildExpenseMetricsLineChartSeries(data, inactiveSeries, isLight),
+    [data, inactiveSeries, isLight]
+  );
 
   return {
     selectedGranularity,
